feat: typed deserialization
Currently the outter deserialization returns a `DecodeType` that we
expect users to cast to their own type `T`. This change modifies the
interface from `deserialize(...): DecodeType` to an automatic casting
`deserialize<T>(...): T`.

For JS users this makes no difference, since there are no types involved
anyway.

For TS users this is a **breaking** change, since `deserialize` will now
return an `unknown` type instead of a `DecodeType` by default if no `T`
is given. However, we would expect users to always be want to cast to `T`.