fix: ut8-string encode/decode
There was a bug by which we were storing the strings as unicode
bytes instead of utf8 bytes. This was a bug since the specification
clearly says that the encoding must be utf8.

This commit fixes such bug using the TextEncode / TextDecode tools,
which are widely supported by modern browsers and node versions.